refactor(models): land AttackPlan in models/attack.py, not the shared dump
models.py grew lumpier when AttackPlan / AttackPlanItem were added to it on
this branch - the file already holds programme, recon, findings, disclosure,
and runtime concerns; piling agent-handoff payloads on top is the same
"split the file" smell pylint enforces at 500 lines (max-module-lines), just
caught earlier by the concerns mixing rather than the line count.

Convert models.py to a package (models/__init__.py) so the new code can land
in models/attack.py from day one. Existing classes stay where they are; no
back-compat re-export from the package root - the four importers that touch
AttackPlan are all on this branch, so they migrate to
"from models.attack import ..." directly. Future per-domain splits
(programme, recon, findings, ...) are deliberately out of scope.
