+++ models/__init__.py
@@ -177,27 +177,6 @@ class ReconResult(BaseModel):
     completed_at: datetime = Field(default_factory=lambda: datetime.now(UTC))
 
 
-# Attack plan (Vulnerability Researcher -> Penetration Tester)
-
-
-class AttackPlanItem(BaseModel):
-    """One probe-target hypothesis from the VR's research pass."""
-
-    probe: str  # CVE id or vulnerability-class name, e.g. "CVE-2022-22965" or "reflected XSS"
-    target: str  # hostname or URL drawn from recon
-    expected_ceiling: Severity  # CRITICAL / HIGH / MEDIUM / LOW the probe could reach
-    rationale: str  # 1-2 sentence "why and what to look for"
-    recon_evidence: list[str]  # references to recon signals that justified this hypothesis
-
-
-class AttackPlan(BaseModel):
-    """The VR's attack plan, handed to the PT and re-read at triage time."""
-
-    programme_handle: str
-    drafted_at: datetime
-    items: list[AttackPlanItem]
-
-
 # Vulnerability findings (Penetration Tester -> Vulnerability Researcher)
