+++ squad/vulnerability_researcher/__init__.py
@@ -9,7 +9,8 @@ from typing import Protocol, cast
 
 from crewai.tools import tool
 
-from models import AttackPlan, AttackPlanItem, Programme, Severity
+from models import Programme, Severity
+from models.attack import AttackPlan, AttackPlanItem
 from squad import SquadMember, read_run_file_tool, read_run_filelist_tool
 from tools import cwe_data, http, owasp_data
 from tools.h1_api import h1
